+++ apps/_shared/ts/helpers/date-helpers.ts
@@ -22,4 +22,33 @@ export default class DateHelpers {
         
         return days_no;
     }
+
+    public static formatDateYMDHIS(d: Date) {
+        const date_parts = [
+            d.getFullYear(),
+            '-',
+            ('00' + (d.getMonth() + 1)).slice(-2),
+            '-',
+            ('00' + d.getDate()).slice(-2),
+            ' ',
+            ('00' + d.getHours()).slice(-2),
+            ':',
+            ('00' + d.getMinutes()).slice(-2),
+            ':',
+            ('00' + d.getSeconds()).slice(-2)
+        ];
+
+        return date_parts.join('');
+    }
+
+    public static setHHMMToDate(d: Date, hhmm: string) {
+        const time_parts = hhmm.split(':');
+        const time_hours = parseInt(time_parts[0]);
+        const time_minutes = parseInt(time_parts[1]);
+        
+        const new_date = new Date(d.getTime());
+        new_date.setHours(time_hours, time_minutes, 0, 0);
+        
+        return new_date;
+    }
 }
