+++ apps/gtfs-rt-status/js/ts/controllers/GTFS_DB_Controller.ts
@@ -1,5 +1,5 @@
 import Progress_Controller from './Progress_Controller'
-import { Date_Helpers } from './../helpers/Date_Helpers' 
+import Date_Helpers from '../_shared/helpers/date-helpers'
 import { URL_Helpers } from '../helpers/URL_Helpers';
 import GTFS_RT_Reporter from './GTFS_RT_Reporter';
